chore: clean all clippy warnings on the wasm32 target
Native was already clippy-clean; the 26 remaining warnings were all on
the wasm32 target (the harness fix in this release made them visible for
the first time via cargo clippy --target wasm32-unknown-unknown).

Mechanical (cargo clippy --fix): io::Error::other, redundant same-type
casts (u32->u32, u64->u64), redundant closures, needless auto-deref,
collapsible if, let-return.

Judgment calls:
- Add is_empty() default to the AsyncRdsInput trait, mirroring the
  native RdsInput trait (the lint correctly flagged the sibling-trait
  asymmetry).
- Rewrite a .min().max() as .clamp() in the wasm buffer-size heuristic.
- #[allow(clippy::too_many_arguments)] on three wasm async streaming
  parser fns, matching the existing convention on their native
  equivalents (the arg count mirrors the native streaming signatures).

clippy --all-targets is now clean on BOTH native and wasm32; 377 native
+ 41 wasm tests pass; fmt clean.

+++ src/wasm/async_chunk_iter.rs
@@ -403,7 +403,7 @@ mod tests {
     async fn async_lazy_character_chunk_iter_reads_values() {
         let mut data = Vec::new();
         for value in ["a", "bbb"] {
-            data.extend_from_slice(&(CHARSXP as u32).to_be_bytes());
+            data.extend_from_slice(&CHARSXP.to_be_bytes());
             data.extend_from_slice(&(value.len() as i32).to_be_bytes());
             data.extend_from_slice(value.as_bytes());
         }
